+++ src/components/library/DocumentTable.tsx
@@ -39,4 +39,3 @@ const DocumentTable = ({ table }) => {
 };
 
 export default DocumentTable;
-
